ci: :green_heart: Add new integration tests suite
