+++ webapp/app/test/integration/proposal.test.ts
@@ -17,9 +17,14 @@ describe("Proposal DB", () => {
 
     expect(proposals[0]).toBeDefined();
     if (!proposals[0]) throw new Error("Proposal not found");
-    expect(proposals[0].calldata.toString().toLowerCase()).toEqual(
+
+    const createdProposal = proposals[0];
+    expect(createdProposal.calldata.toString().toLowerCase()).toEqual(
       calldata.toString().toLowerCase()
     );
+    expect(createdProposal.proposedOn.getTime()).toEqual(proposedOn.getTime());
+    expect(createdProposal.executor.toLowerCase()).toEqual(executor.toLowerCase());
+    expect(createdProposal.transactionHash.toLowerCase()).toEqual(transactionHash.toLowerCase());
   });
 
   it("should ignore proposals with the same external id", async () => {
@@ -34,5 +39,13 @@ describe("Proposal DB", () => {
       .where(eq(proposalsTable.externalId, randomParams.externalId));
 
     expect(proposals).toHaveLength(1);
+  });
+
+  it("should get a list of proposals", async () => {
+    await createRandomProposal();
+    await createRandomProposal();
+    const proposals = await db.select().from(proposalsTable);
+    expect(proposals).toBeDefined();
+    expect(proposals).toBeInstanceOf(Array);
   });
 });
