+++ webapp/config/env.server.ts
@@ -2,12 +2,15 @@ import { createEnv } from "@t3-oss/env-core";
 import { zodHex } from "validate-cli";
 import { z } from "zod";
 
+export const NodeEnvEnum = z.enum(["development", "test", "production"]);
+export type NodeEnv = z.infer<typeof NodeEnvEnum>;
+
 export const env = createEnv({
   server: {
     ALLOW_INDEXING: z.coerce.boolean().default(false),
     DATABASE_URL: z.string(),
     LOG_LEVEL: z.enum(["debug", "info"]).default("info"),
-    NODE_ENV: z.enum(["development", "test", "production"]),
+    NODE_ENV: NodeEnvEnum,
     SERVER_PORT: z.coerce.number().default(3000),
     WALLET_CONNECT_PROJECT_ID: z.string(),
     L1_RPC_URL: z.string(),
